+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 # Changelog
 
-## [Unreleased]
+## [1.1.0] - 2026-06-08
 
 ### Added
 
 - Add .gitattributes file
+
+### Changed
+
+- fix: skip CHANGELOG format check when no pubspec.yaml (TypeScript projects)
 
 ## [1.0.12] - 2025-06-05
 
@@ -96,7 +100,7 @@
 
 - Initial version
 
-[Unreleased]: https://github.com/inlavigo/gg_changelog/compare/1.0.12...HEAD
+[1.1.0]: https://github.com/inlavigo/gg_changelog/compare/1.0.12...1.1.0
 [1.0.12]: https://github.com/inlavigo/gg_changelog/compare/1.0.11...1.0.12
 [1.0.11]: https://github.com/inlavigo/gg_changelog/compare/1.0.10...1.0.11
 [1.0.10]: https://github.com/inlavigo/gg_changelog/compare/1.0.8...1.0.10